+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## In development
 
+- Adding experimental ready delay [#206](https://github.com/nre-learning/antidote-core/pull/206)
 - Convert networkpolicy to use blank namespace selector instead of RFC1918 [#205](https://github.com/nre-learning/antidote-core/pull/205)
 - Upgrade grpc-gateway and add required protobuf options [#204](https://github.com/nre-learning/antidote-core/pull/204)
 - Add (deprecated) 'legacy' image flavor [#202](https://github.com/nre-learning/antidote-core/pull/202)